version ready to be implemented in solps-iter However, temperature solvern ot tested + performances rather poor -> profiling of the code required

+++ modFACE_maths.f90
@@ -17,9 +17,9 @@
      Subroutine DLUDCMP(A,N,INDX,D,CODE)
      IMPLICIT NONE
      integer, parameter :: nmax = 10000
-     real*8, parameter :: tiny = 1.d-30
+     real(DP), parameter :: tiny = 1.d-30
 
-     real*8, intent(inout), dimension(N,N) :: A
+     real(DP), intent(inout), dimension(N,N) :: A
      integer, intent(in) :: N
      integer, intent(out) :: D, CODE
      integer, intent(out), dimension(N) :: INDX
